requirements.txt
flask
nltk
scikit-learn
numpy
PyPDF2
python-docx
gunicorn
